Extract ledger load migration helper
